Amo kategorisering for gjennomforinger
 - readonly på kurstype og spesifisering
 - defaulter til avtalens verdier
