+++ frontend/mr-admin-flate/src/components/avtaler/AvtaleAmoKategoriseringSkjema.tsx
@@ -1,0 +1,202 @@
+import { Checkbox, CheckboxGroup, HGrid, Select, UNSAFE_Combobox } from "@navikt/ds-react";
+import {
+  ForerkortKlasse,
+  InnholdElement,
+  Kurstype,
+  Spesifisering,
+  Toggles,
+} from "mulighetsrommet-api-client";
+import { useFormContext } from "react-hook-form";
+import { useFeatureToggle } from "../../api/features/useFeatureToggle";
+import {
+  forerkortKlasseToString,
+  kurstypeToString,
+  spesifiseringToString,
+} from "../../utils/Utils";
+import { InferredAvtaleSchema } from "../redaksjonelt-innhold/AvtaleSchema";
+
+export function AvtaleAmoKategoriseringSkjema() {
+  const { data: isEnabled } = useFeatureToggle(
+    Toggles.MULIGHETSROMMET_ADMIN_FLATE_ENABLE_GRUPPE_AMO_KATEGORIER,
+  );
+
+  const {
+    setValue,
+    watch,
+    formState: { errors },
+  } = useFormContext<InferredAvtaleSchema>();
+
+  if (!isEnabled) {
+    return null;
+  }
+
+  const kurstype = watch("amoKategorisering.kurstype");
+  const spesifisering = watch("amoKategorisering.spesifisering");
+  const isBransjeSpesifisering =
+    spesifisering &&
+    [
+      Spesifisering.SERVERING_OVERNATTING,
+      Spesifisering.TRANSPORT,
+      Spesifisering.INDUSTRI,
+      Spesifisering.ANDRE_BRANSJER,
+    ].includes(spesifisering);
+  const forerkort = watch("amoKategorisering.forerkort");
+  const norskprove = watch("amoKategorisering.norskprove");
+  const innholdElementer = watch("amoKategorisering.innholdElementer");
+
+  return (
+    <HGrid gap="4" columns={1}>
+      <Select
+        size="small"
+        label="Kurstype"
+        value={kurstype}
+        error={errors?.amoKategorisering?.kurstype?.message}
+        onChange={(type) => {
+          setValue("amoKategorisering.kurstype", type.target.value as Kurstype);
+          setValue("amoKategorisering.norskprove", undefined);
+          setValue("amoKategorisering.spesifisering", undefined);
+          setValue("amoKategorisering.innholdElementer", undefined);
+        }}
+      >
+        <option value={undefined}>Velg kurstype</option>
+        <option value={Kurstype.BRANSJE}>{kurstypeToString(Kurstype.BRANSJE)}</option>
+        <option value={Kurstype.NORSKOPPLAERING}>
+          {kurstypeToString(Kurstype.NORSKOPPLAERING)}
+        </option>
+        <option value={Kurstype.STUDIESPESIALISERING}>
+          {kurstypeToString(Kurstype.STUDIESPESIALISERING)}
+        </option>
+      </Select>
+      {kurstype === Kurstype.BRANSJE && (
+        <Select
+          size="small"
+          label="Spesifisering"
+          hideLabel
+          value={spesifisering}
+          error={errors?.amoKategorisering?.spesifisering?.message}
+          onChange={(type) => {
+            setValue("amoKategorisering.spesifisering", type.target.value as Spesifisering);
+            setValue("amoKategorisering.norskprove", undefined);
+            setValue("amoKategorisering.innholdElementer", undefined);
+          }}
+        >
+          <option value={undefined}>Velg spesifisering</option>
+          <option value={Spesifisering.SERVERING_OVERNATTING}>
+            {spesifiseringToString(Spesifisering.SERVERING_OVERNATTING)}
+          </option>
+          <option value={Spesifisering.TRANSPORT}>
+            {spesifiseringToString(Spesifisering.TRANSPORT)}
+          </option>
+          <option value={Spesifisering.INDUSTRI}>
+            {spesifiseringToString(Spesifisering.INDUSTRI)}
+          </option>
+          <option value={Spesifisering.ANDRE_BRANSJER}>
+            {spesifiseringToString(Spesifisering.ANDRE_BRANSJER)}
+          </option>
+        </Select>
+      )}
+      {kurstype === Kurstype.NORSKOPPLAERING && (
+        <Select
+          size="small"
+          label="Spesifisering"
+          hideLabel
+          value={spesifisering}
+          error={errors?.amoKategorisering?.spesifisering?.message}
+          onChange={(type) => {
+            setValue("amoKategorisering.spesifisering", type.target.value as Spesifisering);
+            setValue("amoKategorisering.norskprove", undefined);
+            setValue("amoKategorisering.innholdElementer", undefined);
+          }}
+        >
+          <option value={undefined}>Velg spesifisering</option>
+          <option value={Spesifisering.NORSKOPPLAERING}>
+            {spesifiseringToString(Spesifisering.NORSKOPPLAERING)}
+          </option>
+          <option value={Spesifisering.FORBEREDENDE_OPPLAERING_FOR_VOKSNE}>
+            {spesifiseringToString(Spesifisering.FORBEREDENDE_OPPLAERING_FOR_VOKSNE)}
+          </option>
+          <option value={Spesifisering.GRUNNLEGGENDE_FERDIGHETER}>
+            {spesifiseringToString(Spesifisering.GRUNNLEGGENDE_FERDIGHETER)}
+          </option>
+        </Select>
+      )}
+      {isBransjeSpesifisering && (
+        <UNSAFE_Combobox
+          clearButton
+          size="small"
+          label="Førerkort"
+          isMultiSelect
+          options={[
+            ForerkortKlasse.A,
+            ForerkortKlasse.A1,
+            ForerkortKlasse.A2,
+            ForerkortKlasse.AM,
+            ForerkortKlasse.AM_147,
+            ForerkortKlasse.B,
+            ForerkortKlasse.B_78,
+            ForerkortKlasse.BE,
+            ForerkortKlasse.C,
+            ForerkortKlasse.C1,
+            ForerkortKlasse.C1E,
+            ForerkortKlasse.CE,
+            ForerkortKlasse.D,
+            ForerkortKlasse.D1,
+            ForerkortKlasse.D1E,
+            ForerkortKlasse.DE,
+            ForerkortKlasse.S,
+            ForerkortKlasse.T,
+          ].map((f) => ({
+            label: forerkortKlasseToString(f),
+            value: f,
+          }))}
+          selectedOptions={
+            forerkort?.map((f) => ({
+              label: forerkortKlasseToString(f),
+              value: f,
+            })) ?? []
+          }
+          onToggleSelected={(option, isSelected) =>
+            isSelected
+              ? setValue("amoKategorisering.forerkort", [
+                  ...(forerkort ?? []),
+                  option as ForerkortKlasse,
+                ])
+              : setValue(
+                  "amoKategorisering.forerkort",
+                  forerkort?.filter((f) => f !== option),
+                )
+          }
+        ></UNSAFE_Combobox>
+      )}
+      {spesifisering === Spesifisering.NORSKOPPLAERING && (
+        <Checkbox
+          value={norskprove ?? false}
+          onChange={() => setValue("amoKategorisering.norskprove", !(norskprove ?? false))}
+          size="small"
+        >
+          Norskprøve
+        </Checkbox>
+      )}
+      {spesifisering && (
+        <CheckboxGroup
+          size="small"
+          legend="Elementer i kurset"
+          onChange={(values) => {
+            setValue("amoKategorisering.innholdElementer", values);
+          }}
+          value={innholdElementer ?? []}
+        >
+          <HGrid columns={2}>
+            <Checkbox value={InnholdElement.GRUNNLEGGENDE_FERDIGHETER}>
+              Grunnleggende ferdigheter
+            </Checkbox>
+            <Checkbox value={InnholdElement.JOBBSOKER_KOMPETANSE}>Jobbsøkerkompetanse</Checkbox>
+            <Checkbox value={InnholdElement.TEORETISK_OPPLAERING}>Teoretisk opplæring</Checkbox>
+            <Checkbox value={InnholdElement.PRAKTISK_OPPLAERING}>Praktisk opplæring</Checkbox>
+            <Checkbox value={InnholdElement.ARBEIDSLIVSKUNNSKAP}>Arbeidslivskunnskap</Checkbox>
+          </HGrid>
+        </CheckboxGroup>
+      )}
+    </HGrid>
+  );
+}
