+++ frontend/mr-admin-flate/src/components/avtaler/AvtaleSkjemaDetaljer.tsx
@@ -2,8 +2,6 @@ import { useAvtaleAdministratorer } from "@/api/ansatt/useAvtaleAdministratorer"
 import { useMigrerteTiltakstyperForAvtaler } from "@/api/tiltakstyper/useMigrerteTiltakstyper";
 import {
   Alert,
-  Checkbox,
-  CheckboxGroup,
   Heading,
   HGrid,
   Loader,
@@ -15,14 +13,10 @@ import {
 import {
   Avtale,
   Avtaletype,
-  ForerkortKlasse,
-  InnholdElement,
-  Kurstype,
   NavAnsatt,
   NavEnhet,
   NavEnhetType,
   Opphav,
-  Spesifisering,
   Tiltakskode,
   TiltakskodeArena,
   Tiltakstype,
@@ -36,7 +30,7 @@ import { MultiValue } from "react-select";
 import { useFeatureToggle } from "../../api/features/useFeatureToggle";
 import { useNusData } from "../../api/nusdata/useNusData";
 import { erAnskaffetTiltak } from "../../utils/tiltakskoder";
-import { addYear, avtaletypeTilTekst, forerkortKlasseToString } from "../../utils/Utils";
+import { addYear, avtaletypeTilTekst } from "../../utils/Utils";
 import { avtaletekster } from "../ledetekster/avtaleLedetekster";
 import { InferredAvtaleSchema } from "../redaksjonelt-innhold/AvtaleSchema";
 import { AdministratorOptions } from "../skjema/AdministratorOptions";
@@ -45,6 +39,7 @@ import { ControlledMultiSelect } from "../skjema/ControlledMultiSelect";
 import { FormGroup } from "../skjema/FormGroup";
 import skjemastyles from "../skjema/Skjema.module.scss";
 import { AvtaleArrangorSkjema } from "./AvtaleArrangorSkjema";
+import { AvtaleAmoKategoriseringSkjema } from "./AvtaleAmoKategoriseringSkjema";
 import { getLokaleUnderenheterAsSelectOptions } from "./AvtaleSkjemaConst";
 
 const minStartdato = new Date(2000, 0, 1);
@@ -189,7 +184,7 @@ export function AvtaleSkjemaDetaljer({ tiltakstyper, ansatt, enheter, avtale }: 
               <AvtaleNUSKategoriVelger />
             ) : null}
             {watch("tiltakstype")?.tiltakskode === Tiltakskode.GRUPPE_ARBEIDSMARKEDSOPPLAERING ? (
-              <AvtaleGruppeAmoKategoriVelger />
+              <AvtaleAmoKategoriseringSkjema />
             ) : null}
           </FormGroup>
 
@@ -414,174 +409,6 @@ function AvtaleNUSKategoriVelger() {
               )
         }
       ></UNSAFE_Combobox>
-    </HGrid>
-  );
-}
-
-function AvtaleGruppeAmoKategoriVelger() {
-  const { data: isEnabled } = useFeatureToggle(
-    Toggles.MULIGHETSROMMET_ADMIN_FLATE_ENABLE_GRUPPE_AMO_KATEGORIER,
-  );
-
-  const {
-    setValue,
-    watch,
-    formState: { errors },
-  } = useFormContext<InferredAvtaleSchema>();
-
-  if (!isEnabled) {
-    return null;
-  }
-
-  const kurstype = watch("amoKategorisering.kurstype");
-  const spesifisering = watch("amoKategorisering.spesifisering");
-  const isBransjeSpesifisering =
-    spesifisering &&
-    [
-      Spesifisering.SERVERING_OVERNATTING,
-      Spesifisering.TRANSPORT,
-      Spesifisering.INDUSTRI,
-      Spesifisering.ANDRE_BRANSJER,
-    ].includes(spesifisering);
-  const forerkort = watch("amoKategorisering.forerkort");
-  const norskprove = watch("amoKategorisering.norskprove");
-  const innholdElementer = watch("amoKategorisering.innholdElementer");
-
-  return (
-    <HGrid gap="4" columns={1}>
-      <Select
-        size="small"
-        label="Kurstype"
-        value={kurstype}
-        error={errors?.amoKategorisering?.kurstype?.message}
-        onChange={(type) => {
-          setValue("amoKategorisering.kurstype", type.target.value as Kurstype);
-          setValue("amoKategorisering.norskprove", undefined);
-          setValue("amoKategorisering.spesifisering", undefined);
-          setValue("amoKategorisering.innholdElementer", undefined);
-        }}
-      >
-        <option value={undefined}>Velg kurstype</option>
-        <option value={Kurstype.BRANSJE}>Bransje-/yrkesrettede kurs</option>
-        <option value={Kurstype.NORSKOPPLAERING}>Norskopplæring/grunnleggende ferdigheter</option>
-        <option value={Kurstype.STUDIESPESIALISERING}>Studiespesialisering</option>
-      </Select>
-      {kurstype === Kurstype.BRANSJE && (
-        <Select
-          size="small"
-          label="Spesifisering"
-          hideLabel
-          value={spesifisering}
-          error={errors?.amoKategorisering?.spesifisering?.message}
-          onChange={(type) => {
-            setValue("amoKategorisering.spesifisering", type.target.value as Spesifisering);
-            setValue("amoKategorisering.norskprove", undefined);
-            setValue("amoKategorisering.innholdElementer", undefined);
-          }}
-        >
-          <option value={undefined}>Velg spesifisering</option>
-          <option value={Spesifisering.SERVERING_OVERNATTING}>Servering/overnatting</option>
-          <option value={Spesifisering.TRANSPORT}>Transport</option>
-          <option value={Spesifisering.INDUSTRI}>Industri</option>
-          <option value={Spesifisering.ANDRE_BRANSJER}>Andre bransjer</option>
-        </Select>
-      )}
-      {kurstype === Kurstype.NORSKOPPLAERING && (
-        <Select
-          size="small"
-          label="Spesifisering"
-          hideLabel
-          value={spesifisering}
-          error={errors?.amoKategorisering?.spesifisering?.message}
-          onChange={(type) => {
-            setValue("amoKategorisering.spesifisering", type.target.value as Spesifisering);
-            setValue("amoKategorisering.norskprove", undefined);
-            setValue("amoKategorisering.innholdElementer", undefined);
-          }}
-        >
-          <option value={undefined}>Velg spesifisering</option>
-          <option value={Spesifisering.NORSKOPPLAERING}>Norskopplæring</option>
-          <option value={Spesifisering.FOV}>Forberedende opplæring for voksne</option>
-          <option value={Spesifisering.GRUNNLEGGENDE_FERDIGHETER}>Grunnleggende ferdigheter</option>
-        </Select>
-      )}
-      {isBransjeSpesifisering && (
-        <UNSAFE_Combobox
-          clearButton
-          size="small"
-          label="Førerkort"
-          isMultiSelect
-          options={[
-            ForerkortKlasse.A,
-            ForerkortKlasse.A1,
-            ForerkortKlasse.A2,
-            ForerkortKlasse.AM,
-            ForerkortKlasse.AM_147,
-            ForerkortKlasse.B,
-            ForerkortKlasse.B_78,
-            ForerkortKlasse.BE,
-            ForerkortKlasse.C,
-            ForerkortKlasse.C1,
-            ForerkortKlasse.C1E,
-            ForerkortKlasse.CE,
-            ForerkortKlasse.D,
-            ForerkortKlasse.D1,
-            ForerkortKlasse.D1E,
-            ForerkortKlasse.DE,
-            ForerkortKlasse.S,
-            ForerkortKlasse.T,
-          ].map((f) => ({
-            label: forerkortKlasseToString(f),
-            value: f,
-          }))}
-          selectedOptions={
-            forerkort?.map((f) => ({
-              label: forerkortKlasseToString(f),
-              value: f,
-            })) ?? []
-          }
-          onToggleSelected={(option, isSelected) =>
-            isSelected
-              ? setValue("amoKategorisering.forerkort", [
-                  ...(forerkort ?? []),
-                  option as ForerkortKlasse,
-                ])
-              : setValue(
-                  "amoKategorisering.forerkort",
-                  forerkort?.filter((f) => f !== option),
-                )
-          }
-        ></UNSAFE_Combobox>
-      )}
-      {spesifisering === Spesifisering.NORSKOPPLAERING && (
-        <Checkbox
-          value={norskprove ?? false}
-          onChange={() => setValue("amoKategorisering.norskprove", !(norskprove ?? false))}
-          size="small"
-        >
-          Norskprøve
-        </Checkbox>
-      )}
-      {spesifisering && (
-        <CheckboxGroup
-          size="small"
-          legend="Elementer i kurset"
-          onChange={(values) => {
-            setValue("amoKategorisering.innholdElementer", values);
-          }}
-          value={innholdElementer ?? []}
-        >
-          <HGrid columns={2}>
-            <Checkbox value={InnholdElement.GRUNNLEGGENDE_FERDIGHETER}>
-              Grunnleggende ferdigheter
-            </Checkbox>
-            <Checkbox value={InnholdElement.JOBBSOKER_KOMPETANSE}>Jobbsøkerkompetanse</Checkbox>
-            <Checkbox value={InnholdElement.TEORETISK_OPPLAERING}>Teoretisk opplæring</Checkbox>
-            <Checkbox value={InnholdElement.PRAKTISK_OPPLAERING}>Praktisk opplæring</Checkbox>
-            <Checkbox value={InnholdElement.ARBEIDSLIVSKUNNSKAP}>Arbeidslivskunnskap</Checkbox>
-          </HGrid>
-        </CheckboxGroup>
-      )}
     </HGrid>
   );
 }
