+++ frontend/mr-admin-flate/src/components/tiltaksgjennomforinger/TiltaksgjennomforingSkjemaContainer.tsx
@@ -95,7 +95,14 @@ export const TiltaksgjennomforingSkjemaContainer = ({
       deltidsprosent: data.deltidsprosent,
       estimertVentetid: data.estimertVentetid ?? null,
       tilgjengeligForArrangorFraOgMedDato: data.tilgjengeligForArrangorFraOgMedDato ?? null,
-      nusData: data.nusData || null,
+      nusData: data.nusData ?? null,
+      amoKategorisering: avtale.amoKategorisering
+        ? {
+            kurstype: avtale.amoKategorisering.kurstype,
+            spesifisering: avtale.amoKategorisering.spesifisering,
+            ...data.amoKategorisering,
+          }
+        : null,
     };
 
     if (
